feat: Tuval's chat window and the atölye exhibit assemble the composer from its parts (#8678)
