+++ .patterns/index.md
@@ -85,6 +85,7 @@ The frontend's component layer (above the fate client data layer). phoenix's UI 
 | [atolye-exhibit-harness.md](./atolye-exhibit-harness.md) | The atölye harness (`apps/web/src/lab/atolye/`): the typed prop-knobs primitive, the `Exhibit`/`defineExhibit` contract, the headless registry, how `ExhibitStage`/`useKnobs`/`PropKnobs` wire a knob change to a prop re-render | Adding an atölye exhibit or building against the harness seam (epic #2473) |
 | [command-palette.md](./command-palette.md) | The shared modal search surface: caller-owned data/navigation, controlled and uncontrolled state, filtering, and the dialog + combobox/listbox keyboard spine | Using or changing `CommandPalette`, or folding a search surface into the `⌘K` contract (ADR 0186) |
 | [agent-chat-pi-rpc.md](./agent-chat-pi-rpc.md) | The atölye-only Pi RPC chat prototype: Vite's local bridge, browser transport, `/` and `@` completion boundaries, and extension UI request round-trips | Adding a Pi-backed agent-chat capability, or changing the local-only broker seam |
+| [agent-chat-compound-parts.md](./agent-chat-compound-parts.md) | The composer's compound members (`AgentChatInput.Root` + the parts): the anatomy in render order, why a host cannot read the Root state it provides, the replace-not-append rule on `.Settings`/`.Toolbar` children, and `.Pickers` as the way to keep the composer's own rows beside a host's | Assembling the composer by hand in a host, or adding a part to it (epic [#8668](https://github.com/kamp-us/phoenix/issues/8668)) |
 | [component-metadata-jsdoc.md](./component-metadata-jsdoc.md) | The per-component metadata JSDoc convention on `packages/design/src/` primitives (descriptive-only): the tag set (`@component`/`@whenToUse`/`@slot`/`@agent`, prop-level JSDoc), the descriptive/normative firewall, TS-compiler-API extractability | Annotating an `@kampus/design` primitive with metadata, or building the doc extractor ([ADR 0194](../.decisions/0194-design-law-jsdoc-firewall.md)) |
 | [markdown-async-block.md](./markdown-async-block.md) | The one `@kampus/design` `Markdown` block that finishes after its first paint (`MermaidBlock`): why the first paint is the real fence and not a placeholder, why the re-measure is the virtualizer's own `ResizeObserver` and not a callback, the cite-a-sanitizer rule for anything arriving as an HTML string, and the two hops that turn a `--token` into a colour a third-party renderer can read | Adding a second async block to `Markdown`, or handing design tokens to a library that paints its own colours ([#8128](https://github.com/kamp-us/phoenix/issues/8128)) |
 | [i18n-catalog.md](./i18n-catalog.md) | The `apps/web` message catalog (`src/i18n/`): per-surface `tr`/`en` files, `CatalogKey` derived from `tr`, the two `satisfies` checks and why they sit in different files, the dynamic `import("./en")` that keeps English off the Turkish path, `useT`'s no-throw default, the brand-noun invariant, the `one`/`other` plural helper | Adding user-facing copy, migrating a surface to `useT`, or adding a catalog key ([ADR 0347](../.decisions/0347-web-copy-behind-i18n-catalog.md)) |
